Added useIsMobile hook for mobile device detection and correction of 61% width in header component input
